requirements.txt
altgraph==0.17.4
dbus-python==1.3.2
packaging==24.2
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
setuptools==75.8.0
